Created One-Page-Website
The HTML website is a foundation for further JavaScript and CSS implementation.
The website has: 1. a navigation section, 2. home section with a citation and an unordered list, a table, a drop down list, and a bookmark; 3. a gallery with embedded pictures and a YouTube video,  a link to another website; 4. a submission form, a newsletter subscription; 5. a footer.