Summary
Created a basic portfolio website with css styling. The website contains basic animations, a video, photos and a text. There is a link to the GitHub profile and a contact form. the website has been uploaded to the web using the FileZilla FTP transfer software and is available online under the domain www.lucydev256.com